Add a juju charm to deploy certification tarmac.
This simply automates the deployment steps described in the top-level
README.md file. An additional undescribed step is sudo configuration
to allow running lxc as the testing user.

Note that creating a YAML file with some access keys is required, see
the README file provided. A bit of configuration on Launchpad itself is
also needed.